+++ svd-encoder/src/addressblock.rs
@@ -4,11 +4,14 @@ impl Encode for crate::svd::AddressBlock {
     type Error = EncodeError;
 
     fn encode(&self) -> Result<Element, EncodeError> {
-        let children = vec![
+        let mut children = vec![
             new_node("offset", format!("0x{:X}", self.offset)),
             new_node("size", format!("0x{:X}", self.size)),
             self.usage.encode_node()?,
         ];
+        if let Some(v) = &self.protection {
+            children.push(v.encode_node()?);
+        };
         let mut elem = Element::new("addressBlock");
         elem.children = children;
         Ok(elem)